move new records state and constants to separate files
