+++ src/core/operationRepo/constants.ts
@@ -1,0 +1,5 @@
+// Reference: OneSignalSDK/onesignal/core/src/main/java/com/onesignal/core/internal/config/ConfigModel.kt
+export const OP_REPO_DEFAULT_FAIL_RETRY_BACKOFF = 15000;
+export const OP_REPO_POST_CREATE_DELAY = 5000;
+export const OP_REPO_EXECUTION_INTERVAL = 5000;
+export const OP_REPO_POST_CREATE_RETRY_UP_TO = 60_000;
